Redirect to new docs location
Please DO NOT merge the branch. It is only pushed
as a backup if we need a redirect again.

I added a clickable link for the users.
I think that is better than a html redirect.
I'm not sure if can confiure a server-side redirect
on github-pages. Anyway, this solution is ok IMHO.